pyproject.toml
[project]
name = "smart-home-server"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "yuya-take" }
]
dependencies = [
    "boto3>=1.34.131",
    "ruff>=0.4.10",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20240622",
    "boto3-stubs>=1.34.131",
    "uvicorn>=0.30.1",
    "fastapi>=0.111.0",
    "asyncio>=3.4.3",
    "slack-sdk>=3.30.0",
    "python-dotenv>=1.0.1",
    "bme680>=2.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
